Feat/chapters (#105)
* add favourite media id to group by

* create service to create generate chapter jobs

* create generate chapters job start

* create example chapters job

* update file names of chapters and thumbnails and standardise timing

* add metadata to media relation to store timestamp

* return chapters from api

* on overwrite remove previous chapters

* sort chapters on created time

* update ws topics and create update for media entity

* update test

+++ internal/ffmpeg/image.go
@@ -21,7 +21,7 @@ func ScaleHeightByWidth(currentHeight, currentWidth, wantedWidth int) int {
 	return int(float32(currentHeight) / float32(currentWidth) * float32(wantedWidth))
 }
 
-func ImageAt(vid string, time int, img string, width, height int) error {
+func ImageAt(vid string, time float64, img string, width, height int) error {
 	if width <= 0 {
 		return fmt.Errorf(ErrNegativeWidth, width)
 	}
